Detect `$::` when accessing variable from different module
In profile classes, it's quite common to access the variables from
component modules that you've just declared.

The code was originally written the way it was to make sure that facts
written as `$::fact_name` didn't get rewritten as `$fact_name` (which
according to the style guide is especially bad).  It was overly cautious
though. We can be certain that any variable with a second set of `::`
isn't a fact, so we don't need to limit the check to variables from our
module only.